Reviese createCalculator
Please read the comments in your code.  Double check in the test specs on how the createCalculator function is invoked, does it need to return a value? Or is it invoked with the new keyword?

What type of value does it return?

Please review and update your test specs, attend office hours or post in discuss if you have additional questions!